+++ Metadata/metadata_functions.cjs
@@ -84,4 +84,34 @@ async function handleExistingOutputFile(rl) {
   return overwrite === 'y' ? '-y' : '-n'; 
 }
 
-module.exports = {createReadline, validateFileExtensions, askQuestion, getContentType, handleExistingOutputFile};
+/**
+ * Cleans the name by removing extra details like year, resolution, and special characters.
+ * @param {string} name - The name to clean.
+ * @returns {string} - The cleaned name.
+ */
+function cleanName(name) {
+  const qualityMarkers = [
+    '2160p', '1080p', '720p', '480p', '360p', 'BluRay', 'WEBRip', 'BRRip',
+    'DVDRip', 'HDRip', 'REPACK', '10bit', 'DUAL-AUDIO', 'KOR-ENG', '6CH', 
+    'x265', 'HEVC-PSA'
+  ];
+
+  // Remove quality markers
+  qualityMarkers.forEach(marker => {
+    const regex = new RegExp(`\\b${marker}\\b`, 'gi');
+    name = name.replace(regex, '');
+  });
+
+  // Remove extra details like year, resolution, etc.
+  name = name.replace(/\d{4}.*$/, '').replace(/[._]/g, ' ');
+
+  // Remove trailing spaces and special characters
+  name = name.trim().replace(/[()[\]{}]+$/, '');
+
+  // Remove multiple spaces
+  name = name.replace(/\s\s+/g, ' ');
+
+  return name;
+}
+
+module.exports = {createReadline, validateFileExtensions, askQuestion, getContentType, handleExistingOutputFile, cleanName};
