+++ package.json
@@ -5,13 +5,14 @@
   "main": "app.js",
   "dependencies": {
     "axios": "^1.3.6",
+    "chai": "^5.1.1",
     "child-process": "^1.0.2",
     "discord-rpc": "^4.0.1",
     "vlc.js": "^3.2.7",
     "vuepress-plugin-open-graph": "^1.9.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha",
     "lint": "eslint --fix .",
     "docs:build": "vuepress build docs",
     "docs:clean-dev": "vuepress dev docs --clean-cache",
@@ -31,6 +32,7 @@
     "@vuepress/theme-default": "^2.0.0-rc.11",
     "eslint": "^8.0.0",
     "eslint-plugin-jsdoc": "^48.2.3",
+    "mocha": "^10.4.0",
     "vue": "^3.4.0",
     "vuepress": "^2.0.0-rc.7"
   }
